extracted `findPath` util to the separated folder, added unit tests
